requirements.txt
# ===== Koto - Python Dependencies =====
# pip install -r requirements.txt

# ── Core (必须) ──
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
requests>=2.31.0
httpx>=0.27.0
psutil>=5.9.0
markdown>=3.5.0
gunicorn>=22.0.0           # 生产环境 WSGI 服务器

# ── Document Processing ──
python-docx>=1.0.0
lxml>=5.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-pptx>=0.6.23
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
jieba>=0.42.1              # 中文分词

# ── Data ──
pandas>=2.0.0
numpy>=1.24.0

# ── Image ──
Pillow>=10.0.0

# ── Scheduling ──
schedule>=1.2.0

# ── Security ──
cryptography>=42.0.0
PyJWT>=2.8.0               # JWT token auth for SaaS

# ── Desktop (Windows only, optional for server mode) ──
# pywebview>=5.0.0
# pystray>=0.19.0
# pyperclip>=1.8.2
# keyboard>=0.13.5
# pyautogui>=0.9.54
# pywin32>=306
# win10toast>=0.9
# wmi>=1.5.1
# uiautomation>=2.0.18

# ── Voice (optional) ──
# SpeechRecognition>=3.10.0
# PyAudio>=0.2.14
# edge-tts>=6.1.0
# vosk>=0.3.45
# flask-sock>=0.7.0

# ── OCR (optional) ──
# pytesseract>=0.3.10

# ── Browser Automation (optional) ──
# selenium>=4.15.0
